requirements.txt
# Production requirements for Render deployment
# Core bot dependencies
python-telegram-bot>=22.3
openai>=1.0.0
python-dotenv==1.0.0

# Web server and async support
aiohttp>=3.9.0
httpx>=0.25.0
gunicorn>=21.2.0
uvicorn[standard]>=0.24.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.40

# Database
SQLAlchemy>=2.0.25
psycopg2-binary==2.9.9
asyncpg>=0.29.0
alembic>=1.12.0

# Security and encryption
cryptography>=41.0.0

# Monitoring and performance
prometheus-client>=0.17.0

# Additional utilities
requests>=2.31.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional ML dependencies (lightweight)
scikit-learn>=1.3.0

# Rate limiting and caching
cachetools>=5.3.0

# Timezone handling
pytz>=2023.3

# JSON handling
orjson>=3.9.0

# Async database drivers
aiopg>=1.4.0

# Environment and configuration
pydantic>=2.0.0

# Logging enhancements
structlog>=23.1.0

# Task scheduling
schedule>=1.2.0